refactor(app): estrai la rete regex+checksum in detectors.py
Sposta iban_ok, piva_ok, cf_ok, luhn_ok, _CF_ODD, DETECTORS e detect_regex
da app.py a un modulo dedicato, senza cambiare una riga di logica.

Il motivo e' lo stesso gia' applicato a pdf_export.py: quel modulo lavora
solo su bytes e dizionario, "quindi e' testabile in isolamento (nessun
import di torch/transformers)". La rete regex ha esattamente lo stesso
bisogno e non lo aveva ancora: importare app.py carica il modello, quindi
oggi non e' possibile scrivere un test sui detector che giri in CI.

app.py ri-esporta i nomi, quindi app.detect_regex continua a esistere
(lo usa src/training/evaluate_targa_stress.py) e il comportamento
pubblico e' invariato.
